Refactor: rename dsv4 moe_ep -> moe, add ep=1 single-card path
moe.py now serves EP world size 1/2/4/8. EP==1 runs a single-card path
(local scatter dispatch_ep1/combine_ep1, no HCCL windows) via moe_ep1 /
moe_ep1_test; EP>=2 keeps the distributed moe / l3_moe path. Naming
convention: unsuffixed = distributed, *_ep1 = single-card.

build_tensor_specs_ep1 reuses the distributed specs (rank-0 projection).
golden_moe_ep1 is a direct local reference: the distributed golden's
route-keyed combine is not bit-equivalent to the single-card token-keyed
combine_ep1 under the random tid2eid fixture (duplicate expert selections),
so the goldens are not interchangeable.

Update decode_layer / prefill_layer imports and doc references.
